+++ test/card_settings_list_picker_test.dart
@@ -24,7 +24,7 @@ void main() {
                   label: label,
                   initialItem: initialValue,
                   items: items,
-                  iconizer: (item) => item.icon,
+                  iconizer: (item) => item?.icon,
                   requiredIndicator: Text(requiredIndicator),
                 )
               ],