+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Releases
 
-## [3.2.0] - 2021-04-07
+## [3.3.0] - 2021-10-25
+
+- Enabled a null or unspecified `initialValue` in `CardSettingsListPicker`
+- Enabled a null or unspecified `initialValue` in `CardSettingsRadioPicker`
+- Added `hintText` property to `CardSettingsNumberPicker` for when there is no initial value
+
+
+## [3.2.0] - 2021-10-24
 
 - upgraded dependencies